Add types for query params and organize param, path, query, and routeMethod types
