+++ src/utilities/path.ts
@@ -1,8 +1,22 @@
-import { Param, ExtractParamsFromPathString, Identity, ParamGetSet, ExtractParamType } from '@/types'
-import { getParamValue, mergeParams, setParamValue } from '@/utilities'
-import { stringHasValue } from '@/utilities/string'
+import { Param, Identity, ReplaceAll, MergeParams, ExtractParamName, ExtractPathParamType } from '@/types'
+import { getParamsForString } from '@/utilities'
 
-type OptionalParam = Param | undefined
+type ParamEnd = '/'
+
+type UnifyParamEnds<
+  TPath extends string
+> = ReplaceAll<ReplaceAll<TPath, '-', ParamEnd>, '_', ParamEnd>
+
+type ExtractParamsFromPathString<
+  TPath extends string,
+  TParams extends Record<string, Param | undefined> = Record<never, never>
+> = UnifyParamEnds<TPath> extends `${infer Path}${ParamEnd}`
+  ? ExtractParamsFromPathString<Path, TParams>
+  : UnifyParamEnds<TPath> extends `${string}:${infer Param}${ParamEnd}${infer Rest}`
+    ? MergeParams<{ [P in ExtractParamName<Param>]: ExtractPathParamType<Param, TParams> }, ExtractParamsFromPathString<Rest, TParams>>
+    : UnifyParamEnds<TPath> extends `${string}:${infer Param}`
+      ? { [P in ExtractParamName<Param>]: [ExtractPathParamType<Param, TParams>] }
+      : Record<never, never>
 
 type PathParams<T extends string> = {
   [K in keyof ExtractParamsFromPathString<T>]?: Param
@@ -16,44 +30,9 @@ export type Path<
   params: Identity<ExtractParamsFromPathString<T, P>>,
 }
 
-function getParam<P extends Record<string, OptionalParam>>(params: P, param: string): Param {
-  return params[param] ?? String
-}
-
 export function path<T extends string, P extends PathParams<T>>(path: T, params: Identity<P>): Path<T, P> {
-  const paramPattern = /:\??([\w]+)(?=\W|$)/g
-  const matches = Array.from(path.matchAll(paramPattern))
-
-  const paramAssignments = matches.map(([match, paramName]) => {
-    const isOptional = match.startsWith(':?')
-    const param = getParam(params, paramName)
-
-    return {
-      [paramName]: [isOptional ? optional(param) : param],
-    }
-  })
-
   return {
     path,
-    params: mergeParams(...paramAssignments) as Path<T, P>['params'],
-  }
-}
-
-export function optional<TParam extends Param>(param: TParam): ParamGetSet<ExtractParamType<TParam> | undefined> {
-  return {
-    get: (value) => {
-      if (!stringHasValue(value)) {
-        return undefined
-      }
-
-      return getParamValue(value, param)
-    },
-    set: (value) => {
-      if (!stringHasValue(value)) {
-        return ''
-      }
-
-      return setParamValue(value, param)
-    },
+    params: getParamsForString(path, params) as Path<T, P>['params'],
   }
 }
